+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/forms/DictTypeForm.java
@@ -1,38 +1,38 @@
 package com.linyun.airline.forms;
 
-import com.linyun.airline.entities.DictTypeEntity;
-import com.uxuexi.core.web.form.SQLParamForm;
-import com.uxuexi.core.db.util.EntityUtil;
+import java.io.Serializable;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 import org.nutz.dao.Cnd;
 import org.nutz.dao.SqlManager;
 import org.nutz.dao.Sqls;
 import org.nutz.dao.sql.Sql;
 
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-import java.io.Serializable;
+import com.linyun.airline.entities.DictTypeEntity;
+import com.uxuexi.core.db.util.EntityUtil;
+import com.uxuexi.core.web.form.SQLParamForm;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class DictTypeForm extends SQLParamForm implements Serializable{
+public class DictTypeForm extends SQLParamForm implements Serializable {
 	private static final long serialVersionUID = 1L;
 	/**主键*/
-	private Integer id;
-	
-	/**字典类别代码*/
+	private long id;
+
+	/**字典类别编码*/
 	private String typeCode;
-	
+
 	/**字典类别名称*/
 	private String typeName;
-	
+
 	/**描述*/
 	private String description;
-	
+
 	/**状态,0-冻结，1-启用，2--删除*/
-	private Integer status;
-	
+	private int status;
+
 	@Override
 	public Sql sql(SqlManager sqlManager) {
 		/**
