+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/forms/SRoleFunctionMapUpdateForm.java
@@ -12,9 +12,9 @@ public class SRoleFunctionMapUpdateForm extends ModForm implements Serializable{
 	private static final long serialVersionUID = 1L;
 		
 	/**角色id*/
-	private Integer roleId;
+	private Long roleId;
 		
 	/**功能id*/
-	private Integer functionId;
+	private Long functionId;
 		
 }
